+++ packages/@n8n/nodes-langchain/nodes/document_loaders/DocumentBinaryInputLoader/DocumentBinaryInputLoader.node.ts
@@ -1,131 +1,8 @@
 /* eslint-disable n8n-nodes-base/node-dirname-against-convention */
-import {
-	IExecuteFunctions,
-	INodeType,
-	INodeTypeDescription,
-	SupplyData,
-	INodeExecutionData,
-	type IBinaryData,
-	NodeOperationError,
-	BINARY_ENCODING,
-} from 'n8n-workflow';
+import type { IExecuteFunctions, INodeType, INodeTypeDescription, SupplyData } from 'n8n-workflow';
 
-import { TextSplitter } from 'langchain/text_splitter';
 import { logWrapper } from '../../../utils/logWrapper';
-import { Document } from 'langchain/document';
-// @ts-ignore
-import { getAndValidateSupplyInput } from '../../../utils/getAndValidateSupplyInput';
-
-import { CSVLoader } from 'langchain/document_loaders/fs/csv';
-import { DocxLoader } from 'langchain/document_loaders/fs/docx';
-import { JSONLoader } from 'langchain/document_loaders/fs/json';
-import { PDFLoader } from 'langchain/document_loaders/fs/pdf';
-import { TextLoader } from 'langchain/document_loaders/fs/text';
-
-import { N8nEPubLoader } from './EpubLoader';
-
-const SUPPORTED_MIME_TYPES = {
-	pdfLoader: ['application/pdf'],
-	csvLoader: ['text/csv'],
-	epubLoader: ['application/epub+zip'],
-	docxLoader: ['application/vnd.openxmlformats-officedocument.wordprocessingml.document'],
-	textLoader: ['text/plain'],
-	jsonLoader: ['application/json'],
-};
-
-export class N8nBinaryLoader {
-	private context: IExecuteFunctions;
-
-	constructor(context: IExecuteFunctions) {
-		this.context = context;
-	}
-
-	async process(items?: INodeExecutionData[]): Promise<Document[]> {
-		const selectedLoader: keyof typeof SUPPORTED_MIME_TYPES = this.context.getNodeParameter(
-			'loader',
-			0,
-		) as keyof typeof SUPPORTED_MIME_TYPES;
-		const binaryDataKey = this.context.getNodeParameter('binaryDataKey', 0) as string;
-		const docs: Document[] = [];
-
-		if (!items) return docs;
-
-		for (let itemIndex = 0; itemIndex < items.length; itemIndex++) {
-			// TODO: Should we support traversing the object to find the binary data?
-			const binaryData = items[itemIndex].binary?.[binaryDataKey] as IBinaryData;
-
-			if (!binaryData) {
-				throw new NodeOperationError(this.context.getNode(), 'No binary data set.');
-			}
-
-			const { data, mimeType } = binaryData;
-
-			if (!Object.values(SUPPORTED_MIME_TYPES).flat().includes(mimeType)) {
-				throw new NodeOperationError(this.context.getNode(), `Unsupported mime type: ${mimeType}`);
-			}
-			if (
-				!SUPPORTED_MIME_TYPES[selectedLoader].includes(mimeType) &&
-				selectedLoader !== 'textLoader'
-			) {
-				throw new NodeOperationError(
-					this.context.getNode(),
-					`Unsupported mime type: ${mimeType} for selected loader: ${selectedLoader}`,
-				);
-			}
-
-			const bufferData = Buffer.from(data, BINARY_ENCODING).buffer;
-			const itemBlob = new Blob([new Uint8Array(bufferData)], { type: mimeType });
-
-			let loader: PDFLoader | CSVLoader | N8nEPubLoader | DocxLoader | TextLoader | JSONLoader;
-			switch (mimeType) {
-				case 'application/pdf':
-					const splitPages = this.context.getNodeParameter('splitPages', 0) as boolean;
-					loader = new PDFLoader(itemBlob, {
-						splitPages,
-					});
-					break;
-				case 'text/csv':
-					const column = this.context.getNodeParameter('column', 0) as string;
-					const separator = this.context.getNodeParameter('separator', 0) as string;
-
-					loader = new CSVLoader(itemBlob, {
-						column,
-						separator,
-					});
-					break;
-				case 'application/epub+zip':
-					loader = new N8nEPubLoader(Buffer.from(bufferData));
-					break;
-				case 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
-					loader = new DocxLoader(itemBlob);
-					break;
-				case 'text/plain':
-					loader = new TextLoader(itemBlob);
-					break;
-				case 'application/json':
-					const pointers = this.context.getNodeParameter('pointers', 0) as string;
-					const pointersArray = pointers.split(',').map((pointer) => pointer.trim());
-					loader = new JSONLoader(itemBlob, pointersArray);
-					break;
-				default:
-					throw new NodeOperationError(
-						this.context.getNode(),
-						`Unsupported mime type: ${mimeType}`,
-					);
-			}
-
-			const textSplitter = (await getAndValidateSupplyInput(this.context, 'textSplitter')) as
-				| TextSplitter
-				| undefined;
-			const loadedDoc = textSplitter
-				? await loader.loadAndSplit(textSplitter)
-				: await loader.load();
-
-			docs.push(...loadedDoc);
-		}
-		return docs;
-	}
-}
+import { N8nBinaryLoader } from '../../../utils/N8nBinaryLoader';
 
 export class DocumentBinaryInputLoader implements INodeType {
 	description: INodeTypeDescription = {
@@ -161,19 +38,9 @@ export class DocumentBinaryInputLoader implements INodeType {
 				required: true,
 				options: [
 					{
-						name: 'PDF Loader',
-						value: 'pdfLoader',
-						description: 'Load PDF documents',
-					},
-					{
 						name: 'CSV Loader',
 						value: 'csvLoader',
 						description: 'Load CSV files',
-					},
-					{
-						name: 'EPub Loader',
-						value: 'epubLoader',
-						description: 'Load EPub files',
 					},
 					{
 						name: 'Docx Loader',
@@ -181,19 +48,29 @@ export class DocumentBinaryInputLoader implements INodeType {
 						description: 'Load Docx documents',
 					},
 					{
-						name: 'Text Loader',
-						value: 'textLoader',
-						description: 'Load plain text files',
+						name: 'EPub Loader',
+						value: 'epubLoader',
+						description: 'Load EPub files',
 					},
 					{
 						name: 'JSON Loader',
 						value: 'jsonLoader',
 						description: 'Load JSON files',
 					},
+					{
+						name: 'PDF Loader',
+						value: 'pdfLoader',
+						description: 'Load PDF documents',
+					},
+					{
+						name: 'Text Loader',
+						value: 'textLoader',
+						description: 'Load plain text files',
+					},
 				],
 			},
 			{
-				displayName: 'Binary data key',
+				displayName: 'Binary Data Key',
 				name: 'binaryDataKey',
 				type: 'string',
 				default: 'data',
@@ -202,8 +79,8 @@ export class DocumentBinaryInputLoader implements INodeType {
 			},
 			// PDF Only Fields
 			{
-				name: 'splitPages',
 				displayName: 'Split Pages',
+				name: 'splitPages',
 				type: 'boolean',
 				default: true,
 				displayOptions: {
@@ -214,8 +91,8 @@ export class DocumentBinaryInputLoader implements INodeType {
 			},
 			// CSV Only Fields
 			{
-				name: 'column',
 				displayName: 'Column',
+				name: 'column',
 				type: 'string',
 				default: '',
 				description: 'Column to extract from CSV',
@@ -226,8 +103,8 @@ export class DocumentBinaryInputLoader implements INodeType {
 				},
 			},
 			{
-				name: 'separator',
 				displayName: 'Separator',
+				name: 'separator',
 				type: 'string',
 				description: 'Separator to use for CSV',
 				default: ',',
@@ -239,8 +116,8 @@ export class DocumentBinaryInputLoader implements INodeType {
 			},
 			// JSON Only Fields
 			{
-				name: 'pointers',
 				displayName: 'Pointers',
+				name: 'pointers',
 				type: 'string',
 				default: '',
 				description: 'Pointers to extract from JSON, e.g. "/text" or "/text, /meta/title"',
@@ -258,7 +135,6 @@ export class DocumentBinaryInputLoader implements INodeType {
 		const processor = new N8nBinaryLoader(this);
 
 		return {
-			// @ts-ignore
 			response: logWrapper(processor, this),
 		};
 	}
