+++ packages/@n8n/nodes-langchain/nodes/embeddings/EmbeddingsOpenAI/EmbeddingsOpenAI.node.ts
@@ -6,7 +6,7 @@ import { logWrapper } from '../../../utils/logWrapper';
 export class EmbeddingsOpenAI implements INodeType {
 	description: INodeTypeDescription = {
 		displayName: 'Embeddings OpenAI',
-		name: 'embeddingsOpenAI',
+		name: 'embeddingsOpenAi',
 		icon: 'file:openAi.svg',
 		credentials: [
 			{
@@ -19,7 +19,6 @@ export class EmbeddingsOpenAI implements INodeType {
 		description: 'Use Embeddings OpenAI',
 		defaults: {
 			name: 'Embeddings OpenAI',
-			color: '#412080',
 		},
 
 		codex: {
