+++ packages/@n8n/nodes-langchain/nodes/llms/LMOpenAi/LMOpenAi.node.ts
@@ -1,8 +1,8 @@
 /* eslint-disable n8n-nodes-base/node-dirname-against-convention */
 import type { IExecuteFunctions, INodeType, INodeTypeDescription, SupplyData } from 'n8n-workflow';
 
-import { logWrapper } from '../../../utils/logWrapper';
 import { OpenAI } from 'langchain';
+import { logWrapper } from '../../../utils/logWrapper';
 
 export class LMOpenAi implements INodeType {
 	description: INodeTypeDescription = {
