+++ packages/@n8n/nodes-langchain/nodes/text_splitters/TextSplitterCharacterTextSplitter/TextSplitterCharacterTextSplitter.node.ts
@@ -1,6 +1,7 @@
 /* eslint-disable n8n-nodes-base/node-dirname-against-convention */
 import type { IExecuteFunctions, INodeType, INodeTypeDescription, SupplyData } from 'n8n-workflow';
-import { CharacterTextSplitter, CharacterTextSplitterParams } from 'langchain/text_splitter';
+import type { CharacterTextSplitterParams } from 'langchain/text_splitter';
+import { CharacterTextSplitter } from 'langchain/text_splitter';
 import { logWrapper } from '../../../utils/logWrapper';
 
 export class TextSplitterCharacterTextSplitter implements INodeType {
