+++ packages/@n8n/nodes-langchain/package.json
@@ -21,8 +21,8 @@
     "clean": "rimraf dist .turbo",
     "dev": "tsc --watch",
     "format": "prettier nodes credentials --write",
-    "lint": "eslint nodes credentials package.json",
-    "lintfix": "eslint nodes credentials package.json --fix",
+    "lint": "eslint nodes credentials",
+    "lintfix": "eslint nodes credentials --fix",
     "prepublishOnly": "npm run build && npm run lint -c .eslintrc.prepublish.js nodes credentials package.json",
     "watch": "tsc-watch -p tsconfig.build.json --onCompilationComplete \"tsc-alias -p tsconfig.build.json\" --onSuccess \"pnpm n8n-generate-ui-types\""
   },
@@ -76,29 +76,27 @@
     ]
   },
   "devDependencies": {
-    "@types/express": "^4.17.6",
-    "@types/html-to-text": "^9.0.1",
-    "@types/request-promise-native": "~1.0.15",
-    "@typescript-eslint/parser": "~5.45",
-    "eslint-plugin-n8n-nodes-base": "^1.11.0",
-    "gulp": "^4.0.2",
-    "n8n-core": "*",
-    "n8n-workflow": "*",
-    "prettier": "^2.7.1",
-    "typescript": "~4.8.4"
-  },
-  "dependencies": {
-    "@gxl/epub-parser": "^2.0.4",
-    "@huggingface/inference": "2.6.1",
-    "@pinecone-database/pinecone": "^0.1.6",
-    "@xata.io/cli": "^0.13.4",
-    "cohere-ai": "^6.2.2",
-    "html-to-text": "^9.0.5",
-    "ignore": "^5.2.4",
-    "langchain": "^0.0.136",
-    "lodash": "^4.17.21",
-    "mammoth": "^1.6.0",
-    "n8n-nodes-base": "workspace:*",
-    "pdf-parse": "^1.1.1"
-  }
+		"@types/express": "^4.17.6",
+		"@types/html-to-text": "^9.0.1",
+		"@types/request-promise-native": "~1.0.15",
+		"eslint-plugin-n8n-nodes-base": "^1.16.0",
+		"gulp": "^4.0.2",
+		"n8n-core": "*"
+	},
+	"dependencies": {
+		"@gxl/epub-parser": "^2.0.4",
+		"@huggingface/inference": "2.6.1",
+		"@pinecone-database/pinecone": "^0.1.6",
+		"@xata.io/cli": "^0.13.4",
+		"@xata.io/client": "^0.26.2",
+		"cohere-ai": "^6.2.2",
+		"html-to-text": "^9.0.5",
+		"ignore": "^5.2.4",
+		"langchain": "^0.0.136",
+		"lodash": "^4.17.21",
+		"mammoth": "^1.6.0",
+		"n8n-nodes-base": "workspace:*",
+		"n8n-workflow": "*",
+		"pdf-parse": "^1.1.1"
+	}
 }
